SPR1-1944 Don't cast chunks to ndarray by default
Set `asarray=False` when calling `from_array`, otherwise dask tries
to force all chunks into ndarrays prematurely. Our chunks are either
ndarrays or of type `PlaceholderChunk` for missing data, which we turn
into ndarrays at the appropriate time.

In the corresponding unit test, tell `map_blocks` its dtype (instead of
letting it figure it out) since it is readily available.

+++ katdal/test/test_chunkstore.py
@@ -244,7 +244,7 @@ class ChunkStoreTestBase:
             else:
                 return chunk
 
-        return da.map_blocks(map_blocks_func, array)
+        return da.map_blocks(map_blocks_func, array, dtype=array.dtype)
 
     def test_dask_array_put_parts_get_whole(self):
         # Split big array into quarters along existing chunks and reassemble
